refactor: ProgressBarへのデータ整形をlib/utils/evaluation-format.tsに切り出し
- formatSectionRatesとformatCategoryRatesを定義
- 複数ページで使い回せる汎用的な配置に変更
